+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log
 
+## [v1.4.0] - 2025-08-20
+
+### Added
+
+- Support blue-green deployment
+    - var: is_enable_blue_green_deployment
+    - resource:  aws_lb_target_group.green
+
 ## [v1.3.1] - 2025-04-09
+
 ### Added
 
 - lifecycle create_before_destroy for aws_lb_target_group